+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -3949,7 +3949,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01M20NAGGHXHJA3Y6DTV2XQ292"),
+  id: z.string().default("01M20SCCDRPSJDN4VKBS72PD3J"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() =>