+++ packages/orq-rc/src/models/operations/updatedataset.ts
@@ -182,7 +182,7 @@ export const UpdateDatasetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2026-09-08T14:03:54.711Z",
+    "2026-09-08T15:14:50.397Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {