+++ packages/orq-rc/RELEASES.md
@@ -12169,3 +12169,13 @@ Based on:
 - [typescript v4.15.0-rc.28] packages/orq-rc
 ### Releases
 - [NPM v4.15.0-rc.28] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.28 - packages/orq-rc
+
+## 2026-09-08 15:18:24
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.796.4 (2.935.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.15.0-rc.29] packages/orq-rc
+### Releases
+- [NPM v4.15.0-rc.29] https://www.npmjs.com/package/@orq-ai/node/v/4.15.0-rc.29 - packages/orq-rc